Move proxy functions to hss.h to mirror the device side model
