+++ Kbuild
@@ -1,2 +1,2 @@
 obj-m += hss.o
-hss-objs := hss-main.o hss-usb.o hss-sockets.o hss-backports.o hss-proxy.o hss-packet.o hss-ring.o
+hss-objs := hss-main.o hss-usb.o hss-sockets.o hss-backports.o hss-proxy.o hss-ring.o
